ENH: Add SSAO for volume rendering
This is a starting pipeline for experimenting around
how to modify / re-purpose SSAO of polyData for volume
rendering. The high-level idea is to calculate a depth
buffer result from the volume rendering step using a threshold
such as 0.5 opacity, and feed into the existing SSAO
algorithm to see what results we can get.